+++ modules/articoli/edit.php
@@ -292,58 +292,6 @@ echo '
 		</div>
 	</div>';
 ?>
-<?php
-
-echo '<div class="panel panel-primary">
-<div class="panel-heading">
-    <h3 class="panel-title">'.tr('Prezzo medio acquisto').'</h3>
-</div>
-
-<div class="panel-body">';
-$rs_prezzo_medio = $dbo->fetchOne('SELECT ((SUM(subtotale)-SUM(sconto))/SUM(qta)) AS prezzo FROM co_righe_documenti INNER JOIN co_documenti ON co_righe_documenti.iddocumento = co_documenti.id WHERE co_documenti.idtipodocumento IN (SELECT id FROM co_tipidocumento WHERE dir = \'uscita\')  AND idarticolo='.prepare($id_record));
-$rs_prezzo_min = $dbo->fetchOne('SELECT ((subtotale-sconto)/qta) AS prezzo, co_documenti.data FROM co_righe_documenti INNER JOIN co_documenti ON co_righe_documenti.iddocumento = co_documenti.id WHERE co_documenti.idtipodocumento IN (SELECT id FROM co_tipidocumento WHERE dir = \'uscita\')  AND idarticolo='.prepare($id_record).' ORDER BY ((subtotale-sconto)/qta) ASC');
-$rs_prezzo_max = $dbo->fetchOne('SELECT ((subtotale-sconto)/qta) AS prezzo, co_documenti.data  FROM co_righe_documenti INNER JOIN co_documenti ON co_righe_documenti.iddocumento = co_documenti.id WHERE co_documenti.idtipodocumento IN (SELECT id FROM co_tipidocumento WHERE dir = \'uscita\')  AND idarticolo='.prepare($id_record).' ORDER BY ((subtotale-sconto)/qta) DESC');
-
-if (count($rs_prezzo_min) > 0) {
-    echo '
-    <div class="row">
-        <div class="col-md-12 col-lg-6">
-            <table class="table table-striped table-condensed table-bordered">
-                <tr>
-                    <th>'.tr('Prezzo minimo').'</th>
-                    <th>'.tr('Prezzio medio').'</th>
-                    <th>'.tr('Prezzo massimo').'</th>
-                    <th>'.tr('Oscillazione').'</th>
-                    <th>'.tr('Oscillazione in %').'</th>
-                    <th>'.tr('Andamento prezzo').'</th>
-                </tr>';
-
-    echo '
-                <tr>
-                    <td>'.moneyFormat($rs_prezzo_min['prezzo']).'</td>
-                    <td>'.moneyFormat($rs_prezzo_medio['prezzo']).'</td>
-                    <td>'.moneyFormat($rs_prezzo_max['prezzo']).'</td>
-                    <td>'.moneyFormat($rs_prezzo_max['prezzo'] - $rs_prezzo_min['prezzo']).'</td>
-                    <td>'.Translator::numberToLocale(((($rs_prezzo_max['prezzo'] - $rs_prezzo_min['prezzo']) * 100) / $rs_prezzo_medio['prezzo']), '2').' %</td>
-                    <td>'.((strtotime($rs_prezzo_min['data']) == strtotime($rs_prezzo_max['data'])) ? 'N.D.' : ((strtotime($rs_prezzo_min['data']) < strtotime($rs_prezzo_max['data'])) ? 'in aumento' : 'in diminuzione')).'</td>
-                </tr>';
-
-    echo '
-                </table>
-            </div>
-        </div>';
-} else {
-    echo '
-        <div class="alert alert-info">
-            '.tr('Questo articolo non è mai stato acquistato').'
-        </div>';
-}
-
-    echo '
-		</div>
-	</div>';
-
-?>
 </form>
 
 {( "name": "filelist_and_upload", "id_module": "$id_module$", "id_record": "$id_record$" )}