+++ templates/registro_iva/bottom.php
@@ -56,4 +56,3 @@ echo '
         </tr>
     </tbody>
 </table>';
-